Introduce an "HIST" file of hdf5 format.
(\sim Abinit HIST.nc file)

The HIST.hdf5 file is updated on-the-fly for each new configuration
added to the database.

For now, only scalar observables of the class `CalcRoutineFunction' are
stored in the HIST.dhf5 file.

Two functions are added to utilies.miscellanous to help retrieve all the
information from the HIST.hdf5 file in a post-processing stage.