Add postprocessing/on-the-fly-visualization tools
Save Nconfs in hist file if mbar is on.

Add a class in utilities.plots, named HistPlot, to handle most of the
functionalities involved in plotting quantities contained in the Hist
file.

Add two commands for on-the-fly visualization: i) mlacs plot_thermo, to
plot Mlacs thermodynamic data (currently, observables are : Temperature
& Total Energy, Pressure & Volume).

ii) mlacs plot_neff, to plot the number of effective configurations as a
function of the number of configurations, and a few past
``distributions'' of weights.

Add post-processing examples in mlacs.examples (with and without mbar).
These examples can easily be used to refine post-processing analysis,
when more information is required than the simple output of mlacs
plot_thermo.