Fix AbinitManager bug
Correct bug of AbinitManager when reading some Abinit files such as
*GSR.nc.  More precisely, in *GSR.nc, the "input_string" dataset stores
Abinit input file, but also some weird not-UTF8-encoded footer. The way
it is separated from the beginning of the file (line breaks or line
breaks with blanck space) does not seem to be universal, which caused
the bug.