Add variable / Change netCDF default
Add electonic entropy, when computed by Abinit. Keyword in *HIST.nc is
'e_entropy'.

Change default netCDF format of output files. Now,
ncformat='NETCDF3_CLASSIC'. This change is aimed at facilitating the use
of Agate (which only reads netCDF3-type formats). The consequence is
that now, by default, two netcdf files are produced by MLACS, the
*HIST.nc and the *WEIGHTS.nc. The latter stores the 'weights' dataset,
whose unlimited dimension is different from "time"
(number-of-configurations).  As a reminder, netCDF3 does not accept more
than one unlimited dimension (this is one of the reasons why the netCDF4
format was introduced). By construction, the 'weights' dataset has
either two unlimited dimensions, or only one, but in that case, it must
be different from "time" (number-of-configurations). This is why weights
are saved in a different file if ncformat='NETCDF3_CLASSIC'.